refactor: add classification (M3201A)--inherits-->(SD_AWG)
Add classification of (M3201A)--inherits-->(SD_AWG), which means a new driver class for each Keysight AWG card can be easily added by specifying card-specifics (e.g. number of channels)
